feat (#33): adicionando edição e barra de pesquisa de usuários
